feat: handle refresh token

+++ src/main/java/com/example/hello_sring_boot/security/JwtProperties.java
@@ -15,7 +15,7 @@ public class JwtProperties {
 
     private String secretKey;
 
-    private long expirationMinute;
+    private int expirationMinute;
 
     private String secretKeyRefresh;
